moved data to its own dir
