+++ src/data/AboutusContent.ts
@@ -1,0 +1,38 @@
+type NextText = {
+  title: string;
+  subTitle: string;
+  imagePath: string;
+};
+
+export const content: NextText[] = [
+  {
+    title: "User RSVPs",
+    subTitle:
+      "Claim your spot at the bar—because great matches deserve packed houses.",
+    imagePath: "next1.jpg",
+  },
+  {
+    title: "Merch Store",
+    subTitle:
+      "Wear your colors—scarves, tees, and pint glasses that scream ‘I run with the crew’.",
+    imagePath: "next2.jpg",
+  },
+  {
+    title: "Supporter Leaderboards",
+    subTitle:
+      "Earn stripes for hosting, cheering, and (responsibly) shotgunning.",
+    imagePath: "next3.jpg",
+  },
+  {
+    title: "Community-Submitted Watch Parties",
+    subTitle:
+      "From backyard projectors to dive bar takeovers—powered by fans like you.",
+    imagePath: "next4.jpg",
+  },
+  {
+    title: "Live Match Data",
+    subTitle:
+      "Real-time stats, heatmaps, and fan chatter—all without leaving your stool.",
+    imagePath: "next5.jpg",
+  },
+];
